rename errors module to exceptions
