+++ src/ikea_api/_api.py
@@ -6,7 +6,7 @@ from typing import Any
 from requests import Session
 
 from ikea_api._constants import DEFAULT_HEADERS
-from ikea_api.errors import GraphQLError, IkeaApiError, UnauthorizedError
+from ikea_api.exceptions import GraphQLError, IkeaApiError, UnauthorizedError
 from ikea_api.types import CustomResponse, GraphQLResponse
