+++ src/ikea_api/_endpoints/item_ingka.py
@@ -2,7 +2,7 @@ from typing import Annotated
 
 from ikea_api._api import API
 from ikea_api._constants import Constants, Secrets
-from ikea_api.errors import ItemFetchError
+from ikea_api.exceptions import ItemFetchError
 from ikea_api.types import CustomResponse
