+++ src/ikea_api/_endpoints/item_pip.py
@@ -1,6 +1,6 @@
 from ikea_api._api import API
 from ikea_api._constants import Constants
-from ikea_api.errors import IkeaApiError
+from ikea_api.exceptions import IkeaApiError
 
 
 class PipItem(API):
