+++ src/ikea_api/exceptions.py
