+++ tests/test_api.py
@@ -7,7 +7,7 @@ import responses
 
 from ikea_api._api import API, AuthorizedAPI, GraphQLAPI
 from ikea_api._constants import DEFAULT_HEADERS
-from ikea_api.errors import GraphQLError, IkeaApiError, UnauthorizedError
+from ikea_api.exceptions import GraphQLError, IkeaApiError, UnauthorizedError
 from ikea_api.types import CustomResponse
